backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
environs==9.5.0
marshmallow==3.18.0
packaging==21.3
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
